[Settings/MG]: Small fixes and additions for the settings UI
Refactors to long and adds a limit to maxGlslCacheSize so java is happy
with people putting absurdly large numbers on it.

Refactors and makes the MG settings appear when current profile is using
it.

Changes the defaults to more logical ones that lets MG handle all the
decision making whenever possible.